requirements.txt
adafruit-circuitpython-ads1x15
adafruit-blinka
pyserial
influxdb-client
pyyaml
RPi.GPIO
